+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/alarm/AlarmRecord.java
@@ -68,7 +68,7 @@ public class AlarmRecord extends Record {
     @Column(columnName = NAME, storageOnly = true, length = 512)
     private String name;
     @Column(columnName = ID0, storageOnly = true, length = 512)
-    @BanyanDB.ShardingKey(index = 0)
+    @BanyanDB.SeriesID(index = 0)
     private String id0;
     @Column(columnName = ID1, storageOnly = true)
     private String id1;